+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3.3 Beta 2 (In development)
 
 - Fixed the issue where a value is still assigned to the `result` when an error occurs in calling pplcv
+- Add the model::timingcache parameter to TensorrtTensor to speed up the construction of models with the same structure in llm
 
 ## v0.3.3 Beta 1 (2023.10.31)
 - Add support for tensorrt 9.1